+++ handbook/src/pages/index.js
@@ -155,7 +155,7 @@ function Gitee() {
             className={"furion-log-jiao" + (isDarkTheme ? " dark" : "")}
           ></div>
           <div className="furion-log-number">
-            <div style={{ color: "#1fd898" }}>110,027</div>
+            <div style={{ color: "#1fd898" }}>111,792</div>
             <span className={isDarkTheme ? " dark" : ""}>Downloads</span>
           </div>
         </div>